+++ js/plugins/image-magnifier.js
@@ -14,11 +14,16 @@ var ImageMagnifier = {
     options: {
         width: "100%",
         height: "auto",
-        magnifierSize: 100,
-        magnifierType: "square", // square, circle
+        lensSize: 100,
+        lensType: "square", // square, circle
         magnifierZoom: 2,
         magnifierMode: "glass", // glass, zoom
         magnifierZoomElement: null,
+
+        clsMagnifier: "",
+        clsLens: "",
+        clsZoom: "",
+
         onMagnifierMove: Metro.noop,
         onImageMagnifierCreate: Metro.noop
     },
@@ -56,12 +61,12 @@ var ImageMagnifier = {
         }
 
         if (!Utils.isValue(element.attr("id"))) {
-            element.attr("id", Utils.elementId("image-comparer"));
+            element.attr("id", Utils.elementId("image-magnifier"));
         }
 
         element.addClass("image-magnifier").css({
             width: o.width
-        });
+        }).addClass(o.clsMagnifier);
 
         element_width = element.width();
 
@@ -77,34 +82,35 @@ var ImageMagnifier = {
             height: element_height
         });
 
-        var x = element_width / 2 - o.magnifierSize / 2;
-        var y = element_height / 2 - o.magnifierSize / 2;
+        var x = element_width / 2 - o.lensSize / 2;
+        var y = element_height / 2 - o.lensSize / 2;
 
         if (o.magnifierMode === "glass") {
 
             magnifier = $("<div>").addClass("image-magnifier-glass").appendTo(element);
             magnifier.css({
-                width: o.magnifierSize,
-                height: o.magnifierSize,
-                borderRadius: o.magnifierType !== "circle" ? 0 : "50%",
+                width: o.lensSize,
+                height: o.lensSize,
+                borderRadius: o.lensType !== "circle" ? 0 : "50%",
                 top: y,
                 left: x,
                 backgroundImage: "url(" + image[0].src + ")",
                 backgroundRepeat: "no-repeat",
+                backgroundPosition: "-" + ((x * o.magnifierZoom) - o.lensSize / 4 + 4) + "px -" + ((y * o.magnifierZoom) - o.lensSize / 4 + 4) + "px",
                 backgroundSize: (image[0].width * o.magnifierZoom) + "px " + (image[0].height * o.magnifierZoom) + "px"
-            });
+            }).addClass(o.clsLens);
 
         } else {
 
             magnifier = $("<div>").addClass("image-magnifier-glass").appendTo(element);
             magnifier.css({
-                width: o.magnifierSize,
-                height: o.magnifierSize,
+                width: o.lensSize,
+                height: o.lensSize,
                 borderRadius: 0,
                 borderWidth: 1,
                 top: y,
                 left: x
-            });
+            }).addClass(o.clsLens);
 
             if (!Utils.isValue(o.magnifierZoomElement) || $(o.magnifierZoomElement).length === 0) {
                 this.zoomElement = $("<div>").insertAfter(element);
@@ -114,10 +120,8 @@ var ImageMagnifier = {
 
             var zoom_element_width = magnifier[0].offsetWidth * o.magnifierZoom;
             var zoom_element_height = magnifier[0].offsetHeight * o.magnifierZoom;
-            var cx = zoom_element_width / o.magnifierSize;
-            var cy = zoom_element_height / o.magnifierSize;
-
-            console.log(cx, cy);
+            var cx = zoom_element_width / o.lensSize;
+            var cy = zoom_element_height / o.lensSize;
 
             this.zoomElement.css({
                 width: zoom_element_width,
@@ -126,21 +130,21 @@ var ImageMagnifier = {
                 backgroundRepeat: "no-repeat",
                 backgroundPosition: "-" + (x * cx) + "px -" + (y * cy) + "px",
                 backgroundSize: (image[0].width * cx) + "px " + (image[0].height * cy) + "px"
-            });
+            }).addClass(o.clsZoom);
         }
     },
 
     _createEvents: function(){
-        var that = this, element = this.element, o = this.options;
-        var magnifier = element.find(".image-magnifier-glass");
-        var magnifier_size = magnifier[0].offsetWidth / 2;
+        var element = this.element, o = this.options;
+        var glass = element.find(".image-magnifier-glass");
+        var glass_size = glass[0].offsetWidth / 2;
         var image = element.find("img")[0];
         var zoomElement = this.zoomElement;
         var cx, cy;
 
         if (o.magnifierMode !== "glass") {
-            cx = zoomElement[0].offsetWidth / magnifier_size / 2;
-            cy = zoomElement[0].offsetHeight / magnifier_size / 2;
+            cx = zoomElement[0].offsetWidth / glass_size / 2;
+            cy = zoomElement[0].offsetHeight / glass_size / 2;
 
             console.log(cx, cy);
 
@@ -149,8 +153,8 @@ var ImageMagnifier = {
             });
         }
 
-        element.on(Metro.events.move, function(e){
-            var x, y, pos = Utils.getCursorPosition(image, e);
+        var lens_move = function(pos){
+            var x, y;
             var magic = 4, zoom = parseInt(o.magnifierZoom);
 
             if (o.magnifierMode === "glass") {
@@ -158,35 +162,35 @@ var ImageMagnifier = {
                 x = pos.x;
                 y = pos.y;
 
-                if (x > image.width - (magnifier_size / zoom)) {
-                    x = image.width - (magnifier_size / zoom);
+                if (x > image.width - (glass_size / zoom)) {
+                    x = image.width - (glass_size / zoom);
                 }
-                if (x < magnifier_size / zoom) {
-                    x = magnifier_size / zoom;
+                if (x < glass_size / zoom) {
+                    x = glass_size / zoom;
                 }
-                if (y > image.height - (magnifier_size / zoom)) {
-                    y = image.height - (magnifier_size / zoom);
+                if (y > image.height - (glass_size / zoom)) {
+                    y = image.height - (glass_size / zoom);
                 }
-                if (y < magnifier_size / zoom) {
-                    y = magnifier_size / zoom;
+                if (y < glass_size / zoom) {
+                    y = glass_size / zoom;
                 }
 
-                magnifier.css({
-                    top: y - magnifier_size,
-                    left: x - magnifier_size,
-                    backgroundPosition: "-" + ((x * zoom) - magnifier_size + magic) + "px -" + ((y * zoom) - magnifier_size + magic) + "px"
+                glass.css({
+                    top: y - glass_size,
+                    left: x - glass_size,
+                    backgroundPosition: "-" + ((x * zoom) - glass_size + magic) + "px -" + ((y * zoom) - glass_size + magic) + "px"
                 });
             } else {
 
-                x = pos.x - (magnifier_size);
-                y = pos.y - (magnifier_size);
+                x = pos.x - (glass_size);
+                y = pos.y - (glass_size);
 
-                if (x > image.width - magnifier_size * 2) {x = image.width - magnifier_size * 2;}
+                if (x > image.width - glass_size * 2) {x = image.width - glass_size * 2;}
                 if (x < 0) {x = 0;}
-                if (y > image.height - magnifier_size * 2) {y = image.height - magnifier_size * 2;}
+                if (y > image.height - glass_size * 2) {y = image.height - glass_size * 2;}
                 if (y < 0) {y = 0;}
 
-                magnifier.css({
+                glass.css({
                     top: y,
                     left: x
                 });
@@ -195,10 +199,29 @@ var ImageMagnifier = {
                     backgroundPosition: "-" + (x * cx) + "px -" + (y * cy) + "px"
                 });
             }
+        };
 
-            Utils.exec(o.onMagnifierMove, [pos, magnifier], element[0]);
+        element.on(Metro.events.move, function(e){
+            var pos = Utils.getCursorPosition(image, e);
+
+            lens_move(pos);
+
+            Utils.exec(o.onMagnifierMove, [pos, glass, zoomElement], element[0]);
 
             e.preventDefault();
+        });
+
+        element.on(Metro.events.leave, function(){
+            var x = element.width() / 2 - o.lensSize / 2;
+            var y = element.height() / 2 - o.lensSize / 2;
+
+            glass.animate({
+                top: y, left: x
+            });
+
+            lens_move({
+                x: x + o.lensSize / 2, y: y + o.lensSize / 2
+            });
         });
     },